requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
fastapi-cache2
redis
